Moved emailer class to separate file.
